Bharati_Pokale/Task_5/Task_Batch_5-6/Anish_Kadam/Task 5/Digit Recognition App/requirements.txt
streamlit>=1.30.0
tensorflow-cpu
numpy
pillow
streamlit-drawable-canvas
matplotlib
scipy


